refactor: API 요청 Hook, Type을 apis로 이동
